tests: improve iamaction and api action tests
A convoluted way was created where a stub is created that always returns an error and where the return value is encapsulated in the error message. Given to how error message are build it was tricky to extract them and while this case worked for most actions it did not work for HeadObject.

Given that these unittest just run in one execution environment it is easier to just introduce a global which gets updated by the stub. So after calling the stub the global will contain the created value. So this approach is easier and more reliable.

Since our stub is called by anonymous requests we should not use session data with valid tags as it makes the tests a bit confusing so changed those.