testing: test case presigned sigv4 URLs should route to correct backend bugfix: backend extraction for pre-signed URLS
This adds test scenario's for using pre-signed URLs to perform GetObject API calls against multiple backends. It also fixes a bug since the selection of a backend was done int he proxy function but that causes issues as that is past the cleaning of the intial request. By extracting the backend prior to proxying we allow to identify the correct backend.

refactor: dry for getting proxy endpoints