feature: multi-backend backward compatibility; allow invalid region
When in the old setup with a single S3 backend an invalid region was provided in a request this would just be ignored as the target region would always be mapped to the configured target region.

For the multi-backend setup specifying an invalid region would result in an internal server error. This change introduces a feature flag ENABLE_LEGACY_BEHAVIOR_INVALID_REGION_TO_DEFAULT_REGION which can be set to true to go back to the old behavior by falling back to the default backend (as specified in backend-config.yaml).

If not set the request would be answered with a Bad Request response explaining to the user that an invalid region was returned.

cleanup: legacy environment variable