bugfix: embed IDP nested claims in credentials
Policies that were using nested claims were not working as expected because the nested claims that were asserted by the IDP were not passed on.

This commit adds test coverage for scenario where IDP claims are in a web identity token which gets exchangd for AWS credentials which are then used and evaluated in a policy.

It also covers the case where the tags are not povided by the IDP. In those cases an internal error was thrown but it is a valid case and instead an Access Denied mus be returned.

This is needed to make  #6 actually work