Feature/multiple fqdns for s3 proxy (#18)
This change allows configuration of the proxy to support multiple FQDNs (via FAKES3PP_S3_PROXY_FQDN)

* test: Almost e2e add coverage for ListObjectsV2

* feature: support providing multiple FQDNs for S3 Proxy

For use cases where an S3 proxy is reachable by multiple DNS names it should be possible to be aware of all of them. Given that a DNS name is used to determine whether virtual hosting is used or not.

* tests: multiple FQDNs for s3proxy test env

* test: ListOjbects against alternate FQDN

It should not matter which FQDN should be used. Add a test case that uses the alternative FQDN.

* refactor: cleanup logic to get FQDN proxy

---------

Co-authored-by: Peter Van Bouwel <peter.vanbouwel@vito.be>