feature: context-aware logging
Make sure to allow logging with context to avoid putting the logic of extracting a requestId everywhere.

Split of logging in its own package and have a requestctx package to track information related to a request.

Introduce hook to allow forcing of logging but currently still disabled.

From this time onwards every logging that can be scoped to a HTTP request should use context aware log call to improve troubleshooting capabilities.