cicd: benchmarking via GH actions (#21)
cicd: benchmarking via GH actions

Benchmark the current HEAD as well as the master branch and compare
performance. Store the results as an artifact into the action.

The repository does not allow commenting on issues using the Github token so we cannot comment the outcome on the PR this might be tackled in the future.