refactor: introduce middleware for logging
Introduce a middleware for logging. The goal of this middleware is to make sure access logging is possible so for accuracy it is best to have it as entry point so even before any router.

Because health checking requires special care for logging this middleware also has health checking as responsibilty.  The same middleware is used for sts and s3 components.