refactor: big bang get CLI isolated from implementation
The cmd package was just a 'dump nearly everything' package. One of the problems of this was tight coupling and not allowing for re-usable parts. This refactor is a starting point. By no means perfect and likley there are way too many public objects but it does help to decouple parts and avoid certain duplication of code.