refactor: introduce middlewares (#23)
* refactor: introduce middleware for logging

Introduce a middleware for logging. The goal of this middleware is to make sure access logging is possible so for accuracy it is best to have it as entry point so even before any router.

Because health checking requires special care for logging this middleware also has health checking as responsibilty.  The same middleware is used for sts and s3 components.

* feature: access log add additional information regarding the request

Make sure the bytes sent received and the status code are updated such that the access log shows the values that are to be expected.

* refactor: s3 api action to stringer in s3 api package

Using stringer allows to track operations easily in Request context and it is also very execution efficient.

* feature: use RegisterOperation middleware

While this initially makes the register route functions uglier we can clean it up later in a refactor. Things will become messier and uglier before they become prettier.

* refactor: move session tags and put s3 under aws services

* refactor: big bang get CLI isolated from implementation

The cmd package was just a 'dump nearly everything' package. One of the problems of this was tight coupling and not allowing for re-usable parts. This refactor is a starting point. By no means perfect and likley there are way too many public objects but it does help to decouple parts and avoid certain duplication of code.

* !backendconfig: relative paths

When relative paths are provided in the backendconfig file they should be relative to the config file and not to the working directory. Absolute paths are just treated as is.

Also allow inline credentials to not overcomplicate creation of test files for test setups.

* !refactor: public credentials functions take interface rather than private key as argument

tests: update presign_test's to work post code restructure

refactor: make sure IsPresignedUrlWithValidSignature also communicates information regarding expiry.

* refactor: move authentication to middleware

Introduce AWS error codes in order to allow to create re-usable code and have an interface that is not service specific.

feature: user facing errors. Allow passing error which can be user facing. #19

* refactor: move authorization to a middleware

* lint: avoid underscore for pointer receiver where name is not used.

* bugfix: avoid test parallelism as it will cause issues with sockets (port 8443 already in use)
bugfix: When awaiting health it should actually shut down the server if not healthy after a while.