Feature/extra info debug (#28)
Summary:

 - tls not set for Basic server eventhough info passed
 - tls variables are no longer mandatory
 - clean all non signed headers when checking signature that covers specific headers
 - make sure to send error back when there is an upstream error
 - tests: allow self-signed certificates for the tests
 - misc change: more debug logging

More detailed

* tracing: extra info on header manipulations

* special handling content length

* test: add dummy test for local debugging.

* bugfix: Basic server was never enabling TLS

* bugfix: no conservatism when cleaning for validating signature

When we have to verify a signature and we are cleaning then we should clean everything that is not considered signed.

This change allows the cleaner to take in options to specify the type of cleaning and e make sure that when checking signature it always cleans.

* docker: example docker compose to run locally.

* logging: more detailed debug

Add information when doign debug logging. Including logging related to the aws signing.

* bugfix: allow self-signe certificates in tests also for presigned urls

* config: tls config parameters are no longer mandatory, they can be left empty for non-tls proxying

* bugfix: fail explicitly when upstream error
bugfix: fail explicitly when using unsupported chunked signing