hook: add logic to call corsHandler
This change lays the foundation to allow configuring CORS for requests. Functionally it does not change anything yet as it comes with a dummy implementation that disables CORS which is similar to what happens prior to this change.

Because for advanced CORS implementations the proxy should be able to go to the backend to ask about the Bucket CORS configuration the interface will be called in the proxy handler as there the required info is present. Since Bucket info is needed and available during authorization it will be added to AccessLogInfo and retrieved when needed.